Split data package into model and repository + exclude them from coverage
