+++ src/main/java/ro/unibuc/hello/config/MongoConfig.java
@@ -1,4 +1,4 @@
-package ro.unibuc.hello.data;
+package ro.unibuc.hello.config;
 
 import com.mongodb.ConnectionString;
 import com.mongodb.MongoClientSettings;
@@ -33,6 +33,6 @@ public class MongoConfig extends AbstractMongoClientConfiguration {
 
     @Override
     public Collection<String> getMappingBasePackages() {
-        return Collections.singleton("ro.unibuc.hello.data");
+        return Collections.singleton("ro.unibuc.hello.model");
     }
 }
