+++ src/main/java/ro/unibuc/hello/repository/TodoRepository.java
@@ -1,9 +1,10 @@
-package ro.unibuc.hello.data;
+package ro.unibuc.hello.repository;
 
 import java.util.List;
 
 import org.springframework.data.mongodb.repository.MongoRepository;
 import org.springframework.stereotype.Repository;
+import ro.unibuc.hello.model.TodoEntity;
 
 @Repository
 public interface TodoRepository extends MongoRepository<TodoEntity, String> {
