+++ src/main/java/ro/unibuc/hello/repository/UserRepository.java
@@ -1,9 +1,10 @@
-package ro.unibuc.hello.data;
+package ro.unibuc.hello.repository;
 
 import java.util.Optional;
 
 import org.springframework.data.mongodb.repository.MongoRepository;
 import org.springframework.stereotype.Repository;
+import ro.unibuc.hello.model.UserEntity;
 
 @Repository
 public interface UserRepository extends MongoRepository<UserEntity, String> {
