+++ src/test/java/ro/unibuc/hello/controller/UserControllerIntegrationTest.java
@@ -3,7 +3,7 @@ package ro.unibuc.hello.controller;
 import com.fasterxml.jackson.databind.ObjectMapper;
 
 import ro.unibuc.hello.IntegrationTestBase;
-import ro.unibuc.hello.data.UserRepository;
+import ro.unibuc.hello.repository.UserRepository;
 import ro.unibuc.hello.request.CreateUserRequest;
 
 import org.junit.jupiter.api.*;
