+++ src/test/java/ro/unibuc/hello/service/UserServiceTest.java
@@ -8,8 +8,8 @@ import org.mockito.Mock;
 import org.mockito.MockitoAnnotations;
 import org.springframework.test.context.junit.jupiter.SpringExtension;
 
-import ro.unibuc.hello.data.UserEntity;
-import ro.unibuc.hello.data.UserRepository;
+import ro.unibuc.hello.model.UserEntity;
+import ro.unibuc.hello.repository.UserRepository;
 import ro.unibuc.hello.request.CreateUserRequest;
 import ro.unibuc.hello.response.UserResponse;
 import ro.unibuc.hello.exception.EntityNotFoundException;
